1. 增加部署模块 2.增加excel导入导出 3. 字典功能 4. 统一异常处理 5. 日志异步配置

+++ {{cookiecutter.project_name}}/{{cookiecutter.app_name}}/api/resources/user.py
@@ -1,10 +1,17 @@
-from flask import request
+from flask import request, Blueprint
 from flask_restful import Resource
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_current_user, get_jwt_identity
 
+from {{cookiecutter.app_name}}.commons import db_util, common_fun
+from {{cookiecutter.app_name}}.commons.constants import const
+from {{cookiecutter.app_name}}.commons.constants.return_code import ReturnCode
 from {{cookiecutter.app_name}}.models import User
 from {{cookiecutter.app_name}}.extensions import ma, db
 from {{cookiecutter.app_name}}.commons.pagination import paginate
+
+logger = logging.getLogger(__name__)
+
+user_bp = Blueprint('user', __name__, url_prefix='/api/v1')
 
 
 class UserSchema(ma.ModelSchema):
@@ -90,26 +97,30 @@ class UserResource(Resource):
 
     def get(self, user_id):
         schema = UserSchema()
-        user = User.query.get_or_404(user_id)
-        return {"user": schema.dump(user).data}
+        user = User.query.get(user_id)
+        if user:
+            ret = {"data": schema.dump(user.__dict__)}
+            return dict(ret, **ReturnCode.OK), 200
+        else:
+            return ReturnCode.NOT_FOUND, 200
 
     def put(self, user_id):
+        session = db.session
         schema = UserSchema(partial=True)
-        user = User.query.get_or_404(user_id)
-        user, errors = schema.load(request.json, instance=user)
-        if errors:
-            return errors, 422
-
-        db.session.commit()
-
-        return {"msg": "user updated", "user": schema.dump(user).data}
+        # user = User.query.get(user_id)
+        user = schema.load(request.json)
+        session.merge(user)
+        session.commit()
+        return ReturnCode.OK, 200
 
     def delete(self, user_id):
-        user = User.query.get_or_404(user_id)
-        db.session.delete(user)
-        db.session.commit()
-
-        return {"msg": "user deleted"}
+        user = User.query.get(user_id)
+        if user:
+            db.session.delete(user)
+            db.session.commit()
+        else:
+            return ReturnCode.NOT_FOUND, 200
+        return ReturnCode.OK, 200
 
 
 class UserList(Resource):
@@ -157,15 +168,61 @@ class UserList(Resource):
     def get(self):
         schema = UserSchema(many=True)
         query = User.query
-        return paginate(query, schema)
+        data = {"data": paginate(query, schema)}
+        return dict(data, **ReturnCode.OK), 200
 
     def post(self):
-        schema = UserSchema()
-        user, errors = schema.load(request.json)
-        if errors:
-            return errors, 422
-
+        schema = UserSchema(unknown=True)
+        user = schema.load(request.json)
         db.session.add(user)
         db.session.commit()
+        return ReturnCode.OK, 200
 
-        return {"msg": "user created", "user": schema.dump(user).data}, 201
+
+class UserInfo(Resource):
+    method_decorators = [jwt_required]
+
+    def get(self):
+        user = get_current_user()
+        session = db.session
+        session.refresh(user)
+        # user.roles = [user.roles]
+        schema = UserSchema(unknown=True)
+        if user:
+            ret = {"data": schema.dump(user.__dict__)}
+            return dict(ret, **ReturnCode.OK), 200
+        else:
+            return ReturnCode.USER_NOT_FOUND, 200
+
+
+@user_bp.route('/users/list', methods=['POST'])
+@jwt_required
+def export_list():
+    query = User.query
+    schema = UserSchema(many=True)
+    datas = query.all()
+    is_or_not_map = common_fun.get_dict_map(const.IS_OR_NOT_DICT_KEY)
+    for r in datas:
+        r.is_caiyin = is_or_not_map.get(str(r.is_caiyin), r.is_caiyin)
+    data = {"data": schema.dump(datas)}
+    return dict(data, **ReturnCode.OK), 200
+
+
+{%- if cookiecutter.use_excel == "yes" %}
+@user_bp.route('/users/import', methods=['POST'])
+@jwt_required
+def import_excel():
+    session = db.session
+    column_names = ['id', 'pwd']
+    records = request.get_records(field_name='file', name_columns_by_row=-1, start_row=1, auto_detect_int=False, colnames=column_names)
+    is_or_not_map = common_fun.get_revert_dict_map(const.IS_OR_NOT_DICT_KEY)
+    # for r in records:
+        # r['isCaiyin'] = is_or_not_map.get(str(r['isCaiyin']), r['isCaiyin'])
+    db_util.increment_update(session=session,
+                             data_dict_list=records,
+                             schema_type=UserSchema,
+                             is_src=True,
+                             request_id=get_jwt_identity())
+    session.commit()
+    return ReturnCode.OK, 200
+{%- endif %}
